Architecture redesign.
Double chain is responsible for book keeping the timestamps now. This is
more logical (since it keeps the access time-sorted index list anyway).
Expirator is a functional module. It is supposed to have its own spec,
symbolic model, and integrate with double-chain and double-map.
A bugfix in the double-map implementation;