Proven kkeeper_add_unrelevant, introduce mvc_coherent_key_abscent
Turns out, there is one more precondition for kkeeper_add_unrelevant, requiring
that the added key was abscent in the vector. mvc_coherent_key_abscent (yet
unproven) is designed to supply this assumption.

The test captures the first use of kkeeper_add_unrelevant, which is putting a new
mac address into the dynamic mac-learning table. It ensures that the new
precondition is being satisfied, for example with the lemma introduced.